requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-json-logger==2.0.7
Flask-CORS==4.0.0
gunicorn==21.2.0
